Move migration classes under browse
