+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/MangaItem.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.ui.migration
+package eu.kanade.tachiyomi.ui.browse.migration
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
