+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/MigrationController.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.ui.migration
+package eu.kanade.tachiyomi.ui.browse.migration
 
 import android.view.LayoutInflater
 import android.view.View
