+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/SearchPresenter.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.ui.migration
+package eu.kanade.tachiyomi.ui.browse.migration
 
 import android.os.Bundle
 import com.jakewharton.rxrelay.BehaviorRelay
