+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/ViewState.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.ui.migration
+package eu.kanade.tachiyomi.ui.browse.migration
 
 import eu.kanade.tachiyomi.source.Source
